rename package to inverter
